requirements.txt
websockets
SpeechRecognition
faster_whisper
fastapi[all]
python-dotenv